create seperate folder for routes
